only backup if it exists

+++ lib/ruby/api/system/SystemPreferences.rb
@@ -42,7 +42,9 @@ def get_default_domain()
  end  
     
 def save_preferences
-  File.rename( SysConfig.SystemPreferencesFile,   SysConfig.SystemPreferencesFile + ".bak")
+  if File.exists?(SysConfig.SystemPreferencesFile) == true
+    File.rename( SysConfig.SystemPreferencesFile,   SysConfig.SystemPreferencesFile + ".bak")
+  end
   serialized_object = YAML::dump(@preferences)
   f = File.new(SysConfig.SystemPreferencesFile,File::CREAT|File::TRUNC|File::RDWR, 0644)
   f.puts(serialized_object)
